Develop

+++ backend/src/main/java/kr/mj/gollaba/auth/service/AuthService.java
@@ -1,12 +1,12 @@
 package kr.mj.gollaba.auth.service;
 
-import kr.mj.gollaba.exception.GollabaErrorCode;
-import kr.mj.gollaba.exception.GollabaException;
 import kr.mj.gollaba.auth.JwtTokenProvider;
 import kr.mj.gollaba.auth.PrincipalDetails;
 import kr.mj.gollaba.auth.dto.LoginRequest;
 import kr.mj.gollaba.auth.dto.LoginResponse;
 import kr.mj.gollaba.auth.repository.UserTokenRepository;
+import kr.mj.gollaba.exception.GollabaErrorCode;
+import kr.mj.gollaba.exception.GollabaException;
 import kr.mj.gollaba.user.entity.User;
 import kr.mj.gollaba.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -26,14 +26,14 @@ public class AuthService implements UserDetailsService {
     private final PasswordEncoder passwordEncoder;
 
     public LoginResponse login(LoginRequest request) {
-        User user = userRepository.findByUniqueId(request.getId())
+        User user = userRepository.findByEmail(request.getEmail())
                 .orElseThrow(() -> new GollabaException(GollabaErrorCode.FAIL_LOGIN));
 
         if (!passwordEncoder.matches(request.getPassword(), user.getPassword())) {
             throw new GollabaException(GollabaErrorCode.FAIL_LOGIN);
         }
 
-        String accessToken = jwtTokenProvider.createAccessToken(user.getUniqueId(), user.getNickName());
+        String accessToken = jwtTokenProvider.createAccessToken(user);
         String refreshToken = jwtTokenProvider.createRefreshToken();
 
         return LoginResponse.builder()
@@ -43,15 +43,20 @@ public class AuthService implements UserDetailsService {
     }
 
     public void logout(String refreshToken) {
+        if (!userTokenRepository.existsByRefreshToken(refreshToken)) {
+            throw new GollabaException(GollabaErrorCode.NOT_EXIST_REFRESH_TOKEN);
+        }
+
         userTokenRepository.deleteByRefreshToken(refreshToken);
     }
 
     @Override
-    public UserDetails loadUserByUsername(String uniqueId) throws UsernameNotFoundException {
-        User user = userRepository.findByUniqueId(uniqueId)
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        User user = userRepository.findByEmail(email)
                 .orElseThrow(() -> new GollabaException(GollabaErrorCode.NOT_EXIST_USER_BY_UNIQUE_ID));
 
         return new PrincipalDetails(user);
     }
+
 
 }
